mm/vmstat: add quiet_vmstat_sync function
To implement task isolation we have to be able to stop vmstat worker
before entering task isolation mode. Therefore there should be a
synchronous version of the function that guarantees that the vmstat
worker will not run on the core on return from the function. In commit
f01f17d3705b ("mm, vmstat: make quiet_vmstat lighter") the
quiet_vmstat() function became asynchronous, in the sense that the
vmstat work was still scheduled to run on the core when the function
returned.

Add a quiet_vmstat_sync() function with that semantic.

Signed-off-by: Chris Metcalf <cmetcalf@mellanox.com>
Signed-off-by: Alex Belits <abelits@marvell.com>