mm/vmstat: add vmstat_idle function
Task isolation mode can not be entered while vmstat worker is running or
should run on the given CPU. This function checks to see if a vmstat
worker is not running, and the vmstat diffs don't require an update. The
function is called from the task-isolation code to see if we need to
actually do some work to quiet vmstat.

Signed-off-by: Chris Metcalf <cmetcalf@mellanox.com>
Signed-off-by: Alex Belits <abelits@marvell.com>