task isolation: prevent sched tick from interrupting
In latency sensitive application, the scheduler tick may be regarded
as a source of the interference.

To keep the process from this interrupt, the IPI to make the core to
update the scheduler tick can be disabled via prctl interface.

If the task isolation flag has been set, the scheduler tick can be
ignored until task isolation is disabled.

Signed-off-by: Oscar Shiang <oscar0225@livemail.tw>